+++ core/oltp/engine/storage/multi-version/mv-record-list.cpp
@@ -50,7 +50,7 @@ void* MV_RecordList_Full::get_readable_version(version_t* head,
 
 std::bitset<1> MV_RecordList_Full::get_readable_version(
     const DeltaList& delta_list, uint64_t tid_self, char* write_loc,
-    const std::vector<std::pair<size_t, size_t>>& column_size_offset_pairs,
+    const std::vector<std::pair<uint16_t, uint16_t>>& column_size_offset_pairs,
     const ushort* col_idx, const ushort num_cols) {
   static thread_local std::bitset<1> ret_bitmask("1");
 
@@ -86,7 +86,7 @@ std::bitset<1> MV_RecordList_Full::get_readable_version(
 
 std::bitset<64> MV_RecordList_Partial::get_readable_version(
     const DeltaList& delta_list, uint64_t tid_self, char* write_loc,
-    const std::vector<std::pair<size_t, size_t>>& column_size_offset_pairs,
+    const std::vector<std::pair<uint16_t, uint16_t>>& column_size_offset_pairs,
     const ushort* col_idx, ushort num_cols) {
   auto* delta_list_ptr = (version_chain_t*)(delta_list.ptr());
 
@@ -173,7 +173,7 @@ std::bitset<64> MV_RecordList_Partial::get_readable_version(
 
 std::vector<MV_RecordList_Full::version_t*> MV_RecordList_Full::create_versions(
     uint64_t xid, global_conf::IndexVal* idx_ptr,
-    std::vector<size_t>& attribute_widths, storage::DeltaStore& deltaStore,
+    std::vector<uint16_t>& attribute_widths, storage::DeltaStore& deltaStore,
     ushort partition_id, const ushort* col_idx, short num_cols) {
   size_t ver_size = 0;
   for (auto& attr_w : attribute_widths) {
@@ -191,7 +191,7 @@ std::vector<MV_RecordList_Full::version_t*> MV_RecordList_Full::create_versions(
 std::vector<MV_RecordList_Partial::version_t*>
 MV_RecordList_Partial::create_versions(uint64_t xid,
                                        global_conf::IndexVal* idx_ptr,
-                                       std::vector<size_t>& attribute_widths,
+                                       std::vector<uint16_t>& attribute_widths,
                                        storage::DeltaStore& deltaStore,
                                        ushort partition_id,
                                        const ushort* col_idx, short num_cols) {
